+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog for OpenEstate-IS24-REST
 0.4-SNAPSHOT (not released yet)
 -------------------------------
 
+-   Updated `commons-lang` library to version 3.6.
+-   Add dependency to `commons-text` library in version 1.1.
+
 
 0.3.3 (14 May 2017)
 -------------------